requirements.txt
slack-bolt==1.18.0
supabase==2.3.0
python-dotenv==1.0.0

